Create haskell_LOTTO.hs
A draft for a lottery smart contract working similar to LOTTO, 
where a player needs to guess a random n-letter word. 
All players buy in tickets at a fixed price, and the higher number of the guessed letters the exponentially higher part of the pot.
An Input should be a set of transactions containing a word,
the script automatically handles words with too few or too many letters, minting the tickets, comparing submissions, handing returns for too low or too high transactions and calculating and preparing transactions for payouts. 
It still needs 
* a random generator plug in to make the random word that others attempt to pre-guess.  
* a plug in to blockChain to read incoming information and execute transactions lists

This code is in development, meant only for preview and backup.